Add validation for premature end of JPG images

+++ scripts/remove_invalid_images.py
@@ -1,5 +1,8 @@
 #!/usr/bin/env python3
 import argparse
+from os import SEEK_END
+from pathlib import Path
+import uuid
 
 import cv2
 
@@ -13,23 +16,45 @@ def parse_args():
     return parser.parse_args()
 
 
+def validate_jpg(path):
+    # Check for premature end of JPEG
+    with path.open('rb') as f:
+        try:
+            f.seek(-2, SEEK_END)
+        except OSError:
+            raise RuntimeError('File too small')
+        if f.read(2) != b'\xff\xd9':
+            raise RuntimeError('Premature end of image')
+
+
 def validate_images(dataset_path):
     print('Validating images in dataset...')
     def validate(path):
-        if path.suffix[1:].isnumeric():
-            new_path = 'duplicate{}.{}'.format(path.suffix, path.stem)
+        try:
+            str(path)
+        except UnicodeEncodeError as e:
+            print('Filename has invalid UTF-8 characters')
+            new_path = Path('{}.{}'.format(uuid.uuid4(), path.suffix))
             path.rename(new_path)
             path = new_path
-        image = cv2.imread(str(path))
+        if path.suffix[1:].isnumeric():
+            new_path = Path('duplicate{}.{}'.format(path.suffix, path.stem))
+            path.rename(new_path)
+            path = new_path
+        if path.suffix.lower() not in ('.bmp', '.gif', '.jpeg', '.jpg', '.pjpeg', '.pjpg', '.png', '.tiff'):
+            return
         try:
+            if path.suffix.lower() in ('.jpeg', '.jpg'):
+                validate_jpg(path)
+            image = cv2.imread(str(path))
             if image is None:
-                raise RuntimeError('Cannot read image {}'.format(path))
+                raise RuntimeError('Cannot read image')
             gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
             histogram = sorted(cv2.calcHist([gray], [0], None, [64], [0, 256]) / (gray.shape[0] * gray.shape[1]))
             if histogram[-1] + histogram[-2] >= 0.9:
                 raise RuntimeError('Histogram shows few light variation')
         except RuntimeError as e:
-            print(e)
+            print('{}: {}'.format(path, e))
             path.unlink()
 
     with ConsumerPool(validate) as validate_consumer_pool:
